Split compiler.js into multiple modules
Which should make it easier to understand what is going on
