+++ lib/compiler/ast-to-ops.js
@@ -1,0 +1,205 @@
+// -*- mode: js; indent-tabs-mode: nil; js-basic-offset: 4 -*-
+//
+// This file is part of ThingTalk
+//
+// Copyright 2015-2018 The Board of Trustees of the Leland Stanford Junior University
+//
+// Author: Giovanni Campagna <gcampagn@cs.stanford.edu>
+//
+// See COPYING for details
+"use strict";
+
+const Ast = require('../ast');
+const { NotImplementedError } = require('../errors');
+
+const { PointWiseOp, StreamOp, TableOp, RuleOp } = require('./ops');
+
+// compile a table that is being monitored to a stream
+function compileMonitorTableToOps(table) {
+    if (table.isVarRef ||
+        table.isAggregation ||
+        table.isArgMinMax ||
+        table.isWindow ||
+        table.isTimeSeries ||
+        table.isHistory ||
+        table.isSequence ||
+        table.isAlias)
+        throw new NotImplementedError(table);
+
+    if (table.isInvocation) {
+        // subscribe is optimistic, we still need EdgeNew
+        return new StreamOp.EdgeNew(new StreamOp.InvokeSubscribe(table.invocation, Ast.BooleanExpression.True));
+    } else if (table.isFilter) {
+        return new StreamOp.Filter(
+            compileMonitorTableToOps(table.table),
+            table.filter);
+    } else if (table.isProjection) {
+        // note the "edge new" operation here, because
+        // the projection might cause fewer values to
+        // be new
+        return new StreamOp.EdgeNew(
+            new StreamOp.Map(
+                compileMonitorTableToOps(table.table),
+                new PointWiseOp.Projection(table.args)
+            ));
+    } else if (table.isCompute) {
+        // note the "edge new" operation here, because
+        // the projection might cause fewer values to
+        // be new
+        return new StreamOp.EdgeNew(
+            new StreamOp.Map(
+                compileMonitorTableToOps(table.table),
+                new PointWiseOp.Compute(table.expression)
+            ));
+    } else if (table.isJoin) {
+        if (table.in_params.length === 0) {
+            // if there is no parameter passing, we can individually monitor
+            // the two tables and return the union
+
+            return new StreamOp.EdgeNew(new StreamOp.Union(
+                compileMonitorTableToOps(table.lhs),
+                compileMonitorTableToOps(table.rhs)));
+        } else {
+            // otherwise we need to subscribe to the left hand side, and
+            // every time it fires, create/update a subscription to the
+            // right hand side
+            // this is VERY MESSY
+            // so it's not implemented
+            throw new NotImplementedError(table);
+        }
+    } else {
+        throw new TypeError();
+    }
+}
+
+// compile a TT stream to a stream op and zero or more
+// tableops
+function compileStreamToOps(stream) {
+    if (stream.isVarRef || stream.isAlias)
+        throw new NotImplementedError(stream);
+
+    if (stream.isTimer) {
+        return new StreamOp.Timer(stream.base, stream.interval);
+    } else if (stream.isAtTimer) {
+        return new StreamOp.AtTimer(stream.time);
+    } else if (stream.isMonitor) {
+        return compileMonitorTableToOps(stream.table);
+    } else if (stream.isEdgeNew) {
+        return new StreamOp.EdgeNew(
+            compileStreamToOps(stream.stream));
+    } else if (stream.isEdgeFilter) {
+        return new StreamOp.EdgeFilter(
+            compileStreamToOps(stream.stream),
+            stream.filter);
+    } else if (stream.isFilter) {
+        return new StreamOp.Filter(
+            compileStreamToOps(stream.stream),
+            stream.filter);
+    } else if (stream.isProjection) {
+        return new StreamOp.Map(
+            compileStreamToOps(stream.stream),
+            new PointWiseOp.Projection(stream.args)
+        );
+    } else if (stream.isCompute) {
+        return new StreamOp.Map(
+            compileStreamToOps(stream.stream),
+            new PointWiseOp.Compute(stream.expression)
+        );
+    } else if (stream.isJoin) {
+        return new StreamOp.Join(
+            compileStreamToOps(stream.stream),
+            compileTableToOps(stream.table, stream.in_params)
+        );
+    } else {
+        throw new TypeError();
+    }
+}
+
+function compileTableToOps(table, extra_in_params) {
+    if (table.isVarRef ||
+        table.isAggregation ||
+        table.isArgMinMax ||
+        table.isWindow ||
+        table.isTimeSeries ||
+        table.isHistory ||
+        table.isSequence ||
+        table.isAlias)
+        throw new NotImplementedError(table);
+
+    if (table.isInvocation) {
+        return new TableOp.InvokeGet(table.invocation, extra_in_params, Ast.BooleanExpression.True);
+    } else if (table.isFilter) {
+        return new TableOp.Filter(
+            compileTableToOps(table.table, extra_in_params),
+            table.filter
+        );
+    } else if (table.isProjection) {
+        return new TableOp.Map(
+            compileTableToOps(table.table, extra_in_params),
+            new PointWiseOp.Projection(table.args)
+        );
+    } else if (table.isCompute) {
+        return new TableOp.Map(
+            compileTableToOps(table.table, extra_in_params),
+            new PointWiseOp.Compute(table.expression)
+        );
+    } else if (table.isJoin) {
+        if (table.in_params.length === 0) {
+            return new TableOp.CrossJoin(
+                compileTableToOps(table.lhs, extra_in_params),
+                compileTableToOps(table.rhs, extra_in_params)
+            );
+        } else {
+            let lhs_in_params = [];
+            let rhs_in_params = [];
+            for (let in_param of extra_in_params) {
+                if (in_param.name in table.lhs.schema.inReq ||
+                    in_param.name in table.lhs.schema.inOpt)
+                    lhs_in_params.push(in_param);
+                if (in_param.name in table.rhs.schema.inReq ||
+                    in_param.name in table.rhs.schema.inOpt)
+                    rhs_in_params.push(in_param);
+            }
+
+            return new TableOp.NestedLoopJoin(
+                compileTableToOps(table.lhs, lhs_in_params),
+                compileTableToOps(table.rhs, rhs_in_params.concat(table.in_params))
+            );
+        }
+    } else {
+        throw new TypeError();
+    }
+}
+
+function optimizeStreamOp(streamop) {
+    return streamop;
+}
+/*function optimizeTableOp(tableop) {
+    return tableop;
+}*/
+
+function optimizeLoop(what, optimizer) {
+    let optimized = optimizer(what);
+    if (optimized !== what)
+        return optimizeLoop(optimized, optimizer);
+    else
+        return optimized;
+}
+
+// compile a rule/command statement to a RuleOp
+function compileStatementToOp(statement) {
+    let streamop = null;
+    if (statement.isRule) {
+        streamop = compileStreamToOps(statement.stream);
+    } else if (statement.table) {
+        let tableop = compileTableToOps(statement.table, []);
+        streamop = new StreamOp.Join(StreamOp.Now, tableop);
+    }
+    streamop = optimizeLoop(streamop, optimizeStreamOp);
+
+    return new RuleOp(streamop, statement.actions);
+}
+
+module.exports = {
+    compileStatementToOp
+};
