+++ lib/compiler/jsir.js
@@ -9,8 +9,8 @@
 // See COPYING for details
 "use strict";
 
-const Builtin = require('./builtin');
-const { stringEscape } = require('./escaping');
+const Builtin = require('../builtin');
+const { stringEscape } = require('../escaping');
 
 // A register-based IR for ThingTalk to JS
 // Typed like ThingTalk
