+++ lib/compiler/ops-to-jsir.js
@@ -2,7 +2,7 @@
 //
 // This file is part of ThingTalk
 //
-// Copyright 2015-2016 The Board of Trustees of the Leland Stanford Junior University
+// Copyright 2015-2018 The Board of Trustees of the Leland Stanford Junior University
 //
 // Author: Giovanni Campagna <gcampagn@cs.stanford.edu>
 //
@@ -10,14 +10,14 @@
 "use strict";
 
 const assert = require('assert');
-const adt = require('adt');
 
-const Ast = require('./ast');
-const Grammar = require('./grammar');
-const Type = require('./type');
+const Ast = require('../ast');
+const Type = require('../type');
+const Builtin = require('../builtin');
+const { NotImplementedError } = require('../errors');
+
 const JSIr = require('./jsir');
-const Builtin = require('./builtin');
-const { typeCheckProgram } = require('./typecheck');
+const { Invocation } = require('./output');
 
 function compileEvent(irBuilder, varScope, name) {
     let reg;
@@ -99,316 +99,14 @@ function compileCast(irBuilder, reg, type, toType) {
     return reg;
 }
 
-class Invocation {
-    constructor(selector, channel, type) {
-        this.selector = selector;
-        this.channel = channel;
-        this.type = type;
-    }
-}
-
 function isRemoteSend(fn) {
     return (fn.selector.kind === 'org.thingpedia.builtin.thingengine.remote' || fn.selector.kind.startsWith('__dyn_')) &&
         fn.channel === 'send';
 }
 
-class NotImplementedError extends Error {
-    constructor(construct) {
-        super('NOT IMPLEMENTED: ' + construct);
-    }
-}
-
-// Low-level ThingTalk operations
-// Each ThingTalk AST node can be implemented in terms of these low-level ops
-// Each of these ops can be compiled into JS code individually
-
-// PointWiseOp : operates on each produced tuple
-const PointWiseOp = adt.data({
-    Projection: { args: adt.only(Array) },
-    Compute: { expression: adt.only(Ast.ScalarExpression) }
-});
-
-// StreamOp : operates on streams
-const StreamOp = adt.data(function() {
-    return {
-        Now: null,
-        InvokeSubscribe: {
-            invocation: adt.only(Ast.Invocation),
-            filter: adt.only(Ast.BooleanExpression)
-        },
-        Timer: {
-            base: adt.only(Ast.Value),
-            interval: adt.only(Ast.Value),
-        },
-        AtTimer: {
-            time: adt.only(Ast.Value),
-        },
-        Filter: {
-            stream: adt.only(this),
-            filter: adt.only(Ast.BooleanExpression)
-        },
-        Map: {
-            stream: adt.only(this),
-            op: adt.only(PointWiseOp)
-        },
-        EdgeNew: {
-            stream: adt.only(this)
-        },
-        EdgeFilter: {
-            stream: adt.only(this),
-            filter: adt.only(Ast.BooleanExpression),
-        },
-        Union: {
-            lhs: adt.only(this),
-            rhs: adt.only(this)
-        },
-    };
-});
-
-// TableOp : operates on in-memory table
-const TableOp = adt.data(function() {
-    return {
-        InvokeGet: {
-            invocation: adt.only(Ast.Invocation),
-            extra_in_params: adt.only(Array), // coming from a join
-            filter: adt.only(Ast.BooleanExpression)
-        },
-
-        Filter: {
-            table: adt.only(this),
-            filter: adt.only(Ast.BooleanExpression)
-        },
-        Map: {
-            table: adt.only(this),
-            op: adt.only(PointWiseOp)
-        },
-
-        CrossJoin: {
-            lhs: adt.only(this),
-            rhs: adt.only(this)
-        },
-        NestedLoopJoin: {
-            lhs: adt.only(this),
-            rhs: adt.only(this),
-        }
-    };
-});
-
-StreamOp.type('Join', {
-    // when the stream fires, get the whole table and join it
-    stream: adt.only(StreamOp),
-    table: adt.only(TableOp)
-});
-
-// The overall structure of the rule
-// this reflects the overall "when => get* => do" structure of ThingTalk
-// which is what it optimizes for
-const RuleOp = adt.newtype('RuleOp', {
-    stream: adt.only(StreamOp, null),
-    actions: adt.only(Array)
-});
-
-// compile a table that is being monitored to a stream
-function compileMonitorTableToOps(table) {
-    if (table.isVarRef ||
-        table.isAggregation ||
-        table.isArgMinMax ||
-        table.isWindow ||
-        table.isTimeSeries ||
-        table.isHistory ||
-        table.isSequence ||
-        table.isAlias)
-        throw new NotImplementedError(table);
-
-    if (table.isInvocation) {
-        // subscribe is optimistic, we still need EdgeNew
-        return new StreamOp.EdgeNew(new StreamOp.InvokeSubscribe(table.invocation, Ast.BooleanExpression.True));
-    } else if (table.isFilter) {
-        return new StreamOp.Filter(
-            compileMonitorTableToOps(table.table),
-            table.filter);
-    } else if (table.isProjection) {
-        // note the "edge new" operation here, because
-        // the projection might cause fewer values to
-        // be new
-        return new StreamOp.EdgeNew(
-            new StreamOp.Map(
-                compileMonitorTableToOps(table.table),
-                new PointWiseOp.Projection(table.args)
-            ));
-    } else if (table.isCompute) {
-        // note the "edge new" operation here, because
-        // the projection might cause fewer values to
-        // be new
-        return new StreamOp.EdgeNew(
-            new StreamOp.Map(
-                compileMonitorTableToOps(table.table),
-                new PointWiseOp.Compute(table.expression)
-            ));
-    } else if (table.isJoin) {
-        if (table.in_params.length === 0) {
-            // if there is no parameter passing, we can individually monitor
-            // the two tables and return the union
-
-            return new StreamOp.EdgeNew(new StreamOp.Union(
-                compileMonitorTableToOps(table.lhs),
-                compileMonitorTableToOps(table.rhs)));
-        } else {
-            // otherwise we need to subscribe to the left hand side, and
-            // every time it fires, create/update a subscription to the
-            // right hand side
-            // this is VERY MESSY
-            // so it's not implemented
-            throw new NotImplementedError(table);
-        }
-    } else {
-        throw new TypeError();
-    }
-}
-
-// compile a TT stream to a stream op and zero or more
-// tableops
-function compileStreamToOps(stream) {
-    if (stream.isVarRef || stream.isAlias)
-        throw new NotImplementedError(stream);
-
-    if (stream.isTimer) {
-        return new StreamOp.Timer(stream.base, stream.interval);
-    } else if (stream.isAtTimer) {
-        return new StreamOp.AtTimer(stream.time);
-    } else if (stream.isMonitor) {
-        return compileMonitorTableToOps(stream.table);
-    } else if (stream.isEdgeNew) {
-        return new StreamOp.EdgeNew(
-            compileStreamToOps(stream.stream));
-    } else if (stream.isEdgeFilter) {
-        return new StreamOp.EdgeFilter(
-            compileStreamToOps(stream.stream),
-            stream.filter);
-    } else if (stream.isFilter) {
-        return new StreamOp.Filter(
-            compileStreamToOps(stream.stream),
-            stream.filter);
-    } else if (stream.isProjection) {
-        return new StreamOp.Map(
-            compileStreamToOps(stream.stream),
-            new PointWiseOp.Projection(stream.args)
-        );
-    } else if (stream.isCompute) {
-        return new StreamOp.Map(
-            compileStreamToOps(stream.stream),
-            new PointWiseOp.Compute(stream.expression)
-        );
-    } else if (stream.isJoin) {
-        return new StreamOp.Join(
-            compileStreamToOps(stream.stream),
-            compileTableToOps(stream.table, stream.in_params)
-        );
-    } else {
-        throw new TypeError();
-    }
-}
-
-function compileTableToOps(table, extra_in_params) {
-    if (table.isVarRef ||
-        table.isAggregation ||
-        table.isArgMinMax ||
-        table.isWindow ||
-        table.isTimeSeries ||
-        table.isHistory ||
-        table.isSequence ||
-        table.isAlias)
-        throw new NotImplementedError(table);
-
-    if (table.isInvocation) {
-        return new TableOp.InvokeGet(table.invocation, extra_in_params, Ast.BooleanExpression.True);
-    } else if (table.isFilter) {
-        return new TableOp.Filter(
-            compileTableToOps(table.table, extra_in_params),
-            table.filter
-        );
-    } else if (table.isProjection) {
-        return new TableOp.Map(
-            compileTableToOps(table.table, extra_in_params),
-            new PointWiseOp.Projection(table.args)
-        );
-    } else if (table.isCompute) {
-        return new TableOp.Map(
-            compileTableToOps(table.table, extra_in_params),
-            new PointWiseOp.Compute(table.expression)
-        );
-    } else if (table.isJoin) {
-        if (table.in_params.length === 0) {
-            return new TableOp.CrossJoin(
-                compileTableToOps(table.lhs, extra_in_params),
-                compileTableToOps(table.rhs, extra_in_params)
-            );
-        } else {
-            let lhs_in_params = [];
-            let rhs_in_params = [];
-            for (let in_param of extra_in_params) {
-                if (in_param.name in table.lhs.schema.inReq ||
-                    in_param.name in table.lhs.schema.inOpt)
-                    lhs_in_params.push(in_param);
-                if (in_param.name in table.rhs.schema.inReq ||
-                    in_param.name in table.rhs.schema.inOpt)
-                    rhs_in_params.push(in_param);
-            }
-
-            return new TableOp.NestedLoopJoin(
-                compileTableToOps(table.lhs, lhs_in_params),
-                compileTableToOps(table.rhs, rhs_in_params.concat(table.in_params))
-            );
-        }
-    } else {
-        throw new TypeError();
-    }
-}
-
-function optimizeStreamOp(streamop) {
-    return streamop;
-}
-/*function optimizeTableOp(tableop) {
-    return tableop;
-}*/
-
-function optimizeLoop(what, optimizer) {
-    let optimized = optimizer(what);
-    if (optimized !== what)
-        return optimizeLoop(optimized, optimizer);
-    else
-        return optimized;
-}
-
-// compile a rule/command statement to a RuleOp
-function compileStatementToOp(statement) {
-    let streamop = null;
-    if (statement.isRule) {
-        streamop = compileStreamToOps(statement.stream);
-    } else if (statement.table) {
-        let tableop = compileTableToOps(statement.table, []);
-        streamop = new StreamOp.Join(StreamOp.Now, tableop);
-    }
-    streamop = optimizeLoop(streamop, optimizeStreamOp);
-
-    return new RuleOp(streamop, statement.actions);
-}
-
-class CompiledRule {
-    constructor(hasTrigger, functions, states, code) {
-        this.hasTrigger = hasTrigger;
-        this.functions = functions;
-        this.states = states;
-        this.code = code;
-    }
-}
-
-class RuleCompiler {
-    constructor(compiler, rule, testMode) {
-        this._testMode = testMode;
-        this._compiler = compiler;
-        this._rule = rule;
-        this._irBuilder = new JSIr.IRBuilder();
+module.exports = class OpCompiler {
+    constructor(irBuilder) {
+        this._irBuilder = irBuilder;
 
         this._functions = [];
         this._nextFunction = 0;
@@ -1050,9 +748,7 @@ class RuleCompiler {
         this._irBuilder.popBlock();
     }
 
-    compile() {
-        let ruleop = compileStatementToOp(this._rule);
-
+    compile(ruleop) {
         if (ruleop.stream)
             this._compileStream(ruleop.stream);
 
@@ -1064,68 +760,6 @@ class RuleCompiler {
         for (let action of ruleop.actions)
             this._compileEndOfFlow(action);
 
-        let result = this._testMode ? this._irBuilder.codegen() : this._irBuilder.compile();
-        return new CompiledRule(this._rule.isRule, this._functions, this._nextStateVar, result);
-    }
-}
-
-module.exports = class AppCompiler {
-    constructor(testMode) {
-        this._testMode = testMode;
-
-        this._name = undefined;
-        this._params = {};
-        this._declarations = {};
-        this._classes = {};
-        this._rules = [];
-
-        this._schemaRetriever = null;
-    }
-
-    setSchemaRetriever(schemaRetriever) {
-        this._schemaRetriever = schemaRetriever;
-    }
-
-    get warnings() {
-        return [];
-    }
-
-    get name() {
-        return this._name;
-    }
-
-    get declarations() {
-        return this._declarations;
-    }
-
-    get rules() {
-        return this._rules;
-    }
-
-    compileCode(code) {
-        return this.compileProgram(Grammar.parse(code));
-    }
-
-    verifyProgram(ast) {
-        return typeCheckProgram(ast, this._schemaRetriever).then(() => {
-            ast.classes.forEach((ast) => {
-                this._classes[ast.name] = ast;
-            });
-            ast.declarations.forEach((ast) => {
-                this._declarations[ast.name] = ast;
-            });
-        });
-    }
-
-    compileProgram(ast) {
-        return this.verifyProgram(ast).then(() => {
-            if (ast.declarations.length > 0)
-                throw new NotImplementedError(ast.declarations);
-
-            ast.rules.forEach((stmt) => {
-                let compiler = new RuleCompiler(this, stmt, this._testMode);
-                this._rules.push(compiler.compile());
-            });
-        });
+        return [this._functions, this._nextStateVar];
     }
 };
